refactor: move authentication schemas to auth.util.ts
