+++ src/lib/utils/auth.util.ts
